requirements.txt
flask==3.0.0
chromadb==0.4.22
ollama==0.1.6
langchain==0.1.4
langchain-community==0.0.13
python-dotenv==1.0.0
Pillow>=10.0.0

